Added createOffer
